requirements.txt
altair==5.1.2
annotated-types==0.7.0
ansi2html==1.8.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.1.0
blinker==1.6.3
branca==0.6.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
Django==5.0.7
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.112.2
fastapi-cli==0.0.5
Flask==2.2.5
folium==0.14.0
fonttools==4.53.1
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.0
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.1
mdurl==0.1.2
motor==3.5.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
nest-asyncio==1.5.8
numpy==1.26.0
orjson==3.10.7
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
plyer==2.1.0
protobuf==4.24.4
pyarrow==13.0.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydeck==0.8.1b0
pygame==2.1.2
Pygments==2.16.1
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.2
ratelim==0.1.6
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rich==13.6.0
rpds-py==0.10.4
scikit-learn==1.3.1
scipy==1.11.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.2
streamlit==1.27.2
streamlit-folium==0.15.0
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typer==0.12.5
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
ujson==5.10.0
urllib3==2.0.6
uvicorn==0.30.6
validators==0.22.0
watchdog==3.0.0
watchfiles==0.24.0
websockets==13.0.1
Werkzeug==2.2.3
zipp==3.17.0
